Moved the cli scripts into the cli pkg
