+++ conf/script/src/build_system/cmd/setup/cli/cli_color.py
